Set header "X-Wreq-Final-Request-URI" on Response.
This is, of course, a pretty nasty hack.  Unfortunately, the "run"
function is the only place to make this change (as it's the only place
Network.HTTP.Client actually makes a request), and its ultimate return
type ("Response") is defined in Network.HTTP.Client.Internal, so it
can't be simply appended to.

I think the proper solution would be to parameterize everything that
currently uses "Response" to allow a choice between "Response" and
"HistoriedResponse" to come from the caller.  But this would require a
very large number of changes.  Simple changes, I suppose, but all over
the codebase.

That's not for me to do, I think.  I just wanted to fix
Network.Scraper.State anyway!